pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tax-assistant"
version = "0.1.0"
description = "Collaborative LLM tax copilot MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.34.0",
  "sqlmodel>=0.0.24",
  "python-multipart>=0.0.20",
  "pypdf>=5.4.0",
  "Pillow>=11.1.0",
  "pydantic-settings>=2.10.0",
  "boto3>=1.39.0",
  "pyjwt[crypto]>=2.10.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.28.0"
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
markers = [
  "integration: tests that require live external services or hosted endpoints",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["tax_assistant*"]
